[CS-3312]: Restructure QrCodeScanner folders/components
