+++ cardstack/src/screens/QRScannerScreen/components/index.ts
@@ -1,8 +1,4 @@
 export * from './EmulatorPasteUriButton';
-export * from './QRCodeScanner';
-export * from './QRCodeScannerCrosshair';
-export * from './QRCodeScannerNeedsAuthorization';
 export * from './RequestQRCode';
 export * from './QRScannerContainer';
 export * from './fixture';
-export * from './QRCodeOverlay';
