+++ cardstack/src/screens/QRScannerScreen/pages/QRCodeScanner/components/QRCodeOverlay.tsx
