+++ cardstack/src/screens/QRScannerScreen/pages/QRCodeScanner/components/index.ts
@@ -1,0 +1,3 @@
+export * from './QRCodeOverlay';
+export { default as CameraNotAuthorizedView } from './CameraNotAuthorizedView';
+export { default as BottomIconsSection } from './BottomIconsSection';
