+++ cardstack/src/screens/QRScannerScreen/pages/QRCodeScanner/strings.ts
@@ -1,0 +1,12 @@
+export const strings = {
+  cameraMountError: 'Error mounting camera',
+  scanQRCodeText: 'Scan a QR code to:',
+  icons: {
+    pay: 'Pay',
+    connect: 'Connect',
+  },
+  authorize: {
+    info: 'Scan to pay or connect',
+    cameraBtn: 'Enable camera access',
+  },
+};
