+++ cardstack/src/screens/QRScannerScreen/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as QRCodeScannerPage } from './QRCodeScanner/QRCodeScannerPage';
