+++ cardstack/src/utils/url-utils.ts
@@ -1,3 +1,6 @@
+import { CARDWALLET_SCHEME } from '@cardstack/cardpay-sdk';
+import Url from 'url-parse';
+
 export const isEncodedUri = (uri = ''): boolean =>
   uri !== decodeURIComponent(uri);
 
@@ -7,4 +10,15 @@ export const decodeNestedURI = (uri: string): string => {
   }
 
   return uri;
+};
+
+// convert https:// to `${CARDWALLET_SCHEME}:/`
+export const convertDeepLinkToCardWalletProtocol = (deepLink: string) => {
+  const url = new Url(deepLink);
+
+  if (url.protocol === 'https:') {
+    return `${CARDWALLET_SCHEME}:/${url.pathname}${url.query || ''}`;
+  }
+
+  return deepLink;
 };
